backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database and ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Pydantic for validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.6
argon2-cffi==23.1.0

# AWS S3 and file handling
boto3==1.34.0
botocore==1.34.0
python-magic==0.4.27

# Email
fastapi-mail==1.4.1
jinja2==3.1.2

# Redis for caching and rate limiting  
redis==5.0.1
hiredis==2.2.3

# HTTP client
httpx==0.25.2
requests==2.31.0

# Utilities
python-dotenv==1.0.0
typer==0.9.0
rich==13.7.0
python-slugify==8.0.1

# Monitoring and logging
sentry-sdk[fastapi]==1.38.0
structlog==23.2.0
prometheus-fastapi-instrumentator==6.1.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
faker==20.1.0

# Code quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8
